Remove Release LABEL to simplify the build process
OSBS system will automatically increment release version based on
the last build if the release LABEL is not specififed. As a result,
there is not needed to have release LABEL anymore. This change
would make the build process simpler as we don't need to manually
update the Dockerful to increment release version.

Signed-off-by: Vu Dinh <vdinh@redhat.com>