Improve Ansible idempotency.
Similar to what we just merged for archivist, the role to deploy the app
will no longer destroy the running app and trigger a build. Instead we
maintain the OpenShift objects, update them if necessary, and check if a
start-build is required on each run.

Several parameters in the template which user integers now fail to
process citing a "int32" error on OpenShift 3.7. Suspect related to:
https://github.com/openshift/origin/issues/3946. These parameters were
never customized so to work around they have been removed and just hard
coded. We can update the template and re-run ansible to adjust in
future, which would be required even to reprocess the template.